requirements.txt
attrs==22.2.0
aws-cdk-lib==2.115.0
aws-cdk.asset-awscli-v1==2.2.201
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v5==2.0.38
aws-cdk.asset-node-proxy-agent-v6==2.0.1
bcrypt==4.0.1
boto3==1.26.56
botocore==1.29.56
cattrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
constructs==10.1.229
cryptography==39.0.0
datadog-cdk-constructs-v2==1.12.0
exceptiongroup==1.1.0
greenlet==2.0.2
importlib-resources==6.1.1
jmespath==1.0.1
jsii==1.93.0
paramiko==3.0.0
publication==0.0.3
pycparser==2.21
PyMySQL==1.0.2
PyNaCl==1.5.0
pysftp==0.2.9
python-dateutil==2.8.2
s3transfer==0.6.0
sentry-sdk==1.14.0
six==1.16.0
SQLAlchemy==2.0.5.post1
typeguard==2.13.3
typing_extensions==4.4.0
urllib3==1.26.14
